update: abc232a

+++ cpp/abc232/a/main.cpp
@@ -4,5 +4,5 @@ using namespace std;
 int main() {
     string s;
     cin >> s;
-    cout << (s[0] - '0') * (s[2] - '0') << endl;
+    cout << (int)(s[0] - '0') * (int)(s[2] - '0') << endl;
 }
